Recipe Id data type updated to Long from Int to hold larger ids

+++ app/src/main/java/dev/ranga/recipeexplorer/api/GetRecipeDetailsUseCase.kt
@@ -3,5 +3,5 @@ package dev.ranga.recipeexplorer.api
 import dev.ranga.recipeexplorer.api.model.RecipeDetail
 
 fun interface GetRecipeDetailsUseCase {
-    suspend fun details(id: Int): RecipeDetail
+    suspend fun details(id: Long): RecipeDetail
 }
