+++ app/src/main/java/dev/ranga/recipeexplorer/data/network/RecipeListResponseDto.kt
@@ -5,15 +5,15 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class RecipeListResponseDto(
-    val count: Int,
+    val count: Long,
     val results: List<RecipeDto>
 )
 
 @Serializable
 data class RecipeDto(
-    val id: Int,
+    val id: Long,
     val name: String? = "",
-    @SerialName("total_time_minutes") val totalTimeMinutes: Int? = 0,
+    @SerialName("total_time_minutes") val totalTimeMinutes: Long? = 0,
     @SerialName("thumbnail_url") val thumbnailUrl: String,
     @SerialName("user_ratings") val userRatings: UserRatingsDto? = null,
 )
